+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v0.4 (Unreleased)
 
 - Add "-config" flag for json config file.
+  This enables configuring multiple services
+  and multiple ports per service.
 - Make command optional (just a port forwarder).
 
 # v0.3 (2017-12-09)